feature/14 /Posts関連のディレクトリ構成見直しと、伴う仕様変更
