+++ src/pages/posts/components/panes/ErrorPane.tsx
@@ -1,0 +1,1 @@
+export const ErrorPane = () => <div style={{ color: "red" }}>エラーです</div>;
