+++ src/pages/posts/components/panes/LoadingPane.tsx
@@ -1,0 +1,1 @@
+export const LoadingPane = () => <div style={{ color: "orange" }}>loading...</div>;
